refactor: move general kafka consumer utilities into separate module
